Refactor NewClaimState to be in its own file
